requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
browser-use==0.1.40
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.43
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.13
markdownify==0.14.1
MarkupSafe==3.0.2
monotonic==1.6
multidict==6.1.0
ollama==0.4.7
openai==1.65.4
orjson==3.10.15
packaging==24.2
playwright==1.50.0
posthog==3.19.0
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
